requirements.txt
dnspython==1.16.0
pymongo==3.12.0
python-dotenv==1.1.0
